Split test_check_info.py
Move revhistory tests into test_check_revhistory.py
